requirements.txt
apache-airflow==2.7.1
apache-airflow-providers-apache-spark==3.0.0
pyspark==3.5.0
pandas==2.1.3
pyarrow==13.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
